Refactored hovering logic into a coroutine in PointerEventsController instead of a new monobehaviour
