--- unity-client/Assets/Scripts/MainScripts/DCL/Controllers/InteractionDetectionController/InteractionDetectionController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 31d8abdc7cc724847a3e405981121e19
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
